feat(login): огибающая по потолку фактической популяции — время исхода не называет класс
Отказ на неверном пароле стоил столько, сколько стоит проверка хранимого
значения, и сообщал класс этого значения (ID-PW-1 Р4): 18 пар из 21 на
Ф3-31 расходились в 25–900 раз. Теперь всякий исход после ворот частоты
уходит не раньше потолка — калиброванной при старте стоимости самого
дорогого класса среди лежащих в хранилище и класса ручки, с запасом в
четверть. Ждёт полоса, не проверяющий: ёмкость отпущена до ожидания.

Класс присутствует в хранилище — узнаётся одним проходом по таблице при
старте: класс вычисляет база и отдаёт значение без соли и тела. Класс,
встреченный на чтении и неизвестный огибающей, калибруется тем же входом
(перенос иным процессом): окно оракула — одно обращение, не до перезапуска.
Потолок, класс-потолок и калибровки по поводу — в самоотчёте старта и
рядами. Проба Ф3-31: 21 пара из 21 на двух прогонах, |Δмедиан| ≤ 0,35 мс.

Refs PRO-Robotech/kaname#188

+++ internal/apps/kaname/api/humansession/timing_probe_test.go
@@ -162,7 +162,7 @@ func TestLogin_F3_31_RefusalTimeIsIndistinguishableAcrossCostClasses(t *testing.
 	envelope, err := passwordverify.NewEnvelope(h.verifier, passwordverify.NopEnvelopeObserver{})
 	require.NoError(t, err)
 	population := []domain.PasswordCostClass{
-		costClass(domain.PasswordHashFormatBcrypt, domain.CostParamBcryptCost, bcrypt.MinCost),
+		costClass(domain.PasswordHashFormatBcrypt, domain.CostParamBcryptCost, uint32(bcrypt.MinCost)),
 		costClass(domain.PasswordHashFormatBcrypt, domain.CostParamBcryptCost, 14),
 		argon2Class(65536, 3, 4), argon2Class(131072, 10, 8), argon2Class(32768, 3, 4),
 		{Format: h.hasher.Declared().Format, Params: h.hasher.Declared().Params},
